Add category "source" to integration tests
