+++ cix.sh
@@ -33,6 +33,7 @@ case "$RUN_ACTIVITY" in
       case "$CATEGORY_GROUP" in
         Category1)
           CATEGORY=Category1 && runCategory
+          CATEGORY=source && runCategory
           ;;
 
         Category2)
